Redefinition of newTypes Sector and NbrSectors to usize. Preparing the cache algorythm.

+++ rust_kernel/src/tests/hard_drive_write_bios.rs
@@ -12,8 +12,8 @@ use crate::drivers::storage::BiosInt13h;
 use crate::drivers::storage::{BlockIo, NbrSectors, Sector};
 
 const NB_TESTS: usize = 48;
-const DISK_SECTOR_CAPACITY: u16 = 0x8000;
-const SECTOR_SIZE: u64 = 512;
+const DISK_SECTOR_CAPACITY: usize = 0x8000;
+const SECTOR_SIZE: usize = 512;
 
 #[no_mangle]
 pub extern "C" fn kmain(
@@ -60,10 +60,10 @@ pub extern "C" fn kmain(
     use alloc::vec::Vec;
 
     for nb_test in 0..NB_TESTS {
-        let start_sector = Sector(srand::<u16>(DISK_SECTOR_CAPACITY - 1) as u64);
-        let mut n = srand::<u16>(1024) as u64;
-        if start_sector.0 + n > DISK_SECTOR_CAPACITY as u64 {
-            n = DISK_SECTOR_CAPACITY as u64 - start_sector.0;
+        let start_sector = Sector(srand::<usize>(DISK_SECTOR_CAPACITY - 1));
+        let mut n = srand::<usize>(1024);
+        if start_sector.0 + n > DISK_SECTOR_CAPACITY {
+            n = DISK_SECTOR_CAPACITY - start_sector.0;
         }
         let nbr_sectors = NbrSectors(n);
